Updating dataframe access pattern
This change is an alternative approach to accessing the main dataframe.
Instead of re-reading the csv multiple times, we hold it in memory as a
global variable, and then copy that dataframe in the individual callbacks.

There is are a couple approaches to making this more memory efficient, but
I don't see performance as an issue yet. Something to explore later though
might be leveraging a sqlite or duckdb database file.